+++ wiring/src/spark_wiring_bluetooth_le.cpp
@@ -73,6 +73,6 @@ void BLEClass::setGatewayTargetName(char* name)
     HAL_BLE_Set_Gateway_Target(name);
 }
 
-void BLEClass::beacon(uint32_t major, uint32_t minor, uint8_t *UUID) {
+void BLEClass::beacon(uint16_t major, uint16_t minor, uint8_t *UUID) {
     HAL_BLE_Start_iBeacon(major, minor, UUID);
 };
